requirements.txt
flask==2.3.1
flask-wtf==1.1.1
wtforms-html5==0.6.1
sib_api_v3_sdk==7.6.0
